[nvim/cpp] Fix bug in makeTree()
Fixes a bug in a snippet inserting a custom function
that generates binary trees from a list of values.
The bug in the logic resulted in the generated trees being flipped
on Y-axis (every node that should be a right-side child was instead
a left-side, and vice versa).
This went unnoticed for quite some time ;-)