[nvim][snip] Answer "-42" by default
The default answer used to be "0"; but this is actually a valid answer
to a lot of example testcases. On the other hand, I've virtually never
seen an example testcase where -42 would be a valid answer.

Make the default integer answer to a testcase -42,
so that all the testcases are expected to fail before the actual solution
is provided.